requirements.txt
python-pptx
python-docx
httpx
pyinstaller
